# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: KhelSetu API
  version: 1.0.0
  description: >
    # KhelSetu API Documentation


    ## Platform Overview


    KhelSetu is a multi-tenant sports tournament management platform with **two
    distinct portals**:


    ### 🌐 Public Portal (No Login Required)

    The public portal is for **viewers and fans** who want to watch live
    matches, browse tournaments, and view standings. These endpoints are
    **completely open** — no authentication needed.


    - **Live match scores** — Real-time scores for any match

    - **Tournament browsing** — View all public tournaments

    - **Standings & fixtures** — League tables and match schedules

    - **Team & player profiles** — Roster and player information

    - **Global search** — Search across tournaments, teams, and players


    ### 🔒 SaaS Dashboard (Login Required)

    The dashboard is for **organization members** (admins, scorers, coaches) who
    manage tournaments, teams, players, and live scoring. **All dashboard
    endpoints require JWT authentication** and **organization membership**.


    - **Organization management** — Create/manage organizations, invite members

    - **Tournament management** — Create tournaments, generate fixtures

    - **Team & player management** — Roster management, player profiles

    - **Live scoring** — Real-time event-driven scoring (Football, Cricket,
    Basketball)

    - **Standings** — Auto-calculated league tables

...

...

...


...


...

...
    refresh token (7d)
...
paths:
...
  /api/matches:
    post:
      summary: Create new match
      tags:
        - Matches
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tournament_id
                - home_team_id
                - away_team_id
              properties:
                tournament_id:
                  type: string
                home_team_id:
                  type: string
                away_team_id:
                  type: string
                round_number:
                  type: integer
                group_name:
                  type: string
                venue:
                  type: string
                scheduled_at:
                  type: string
      responses:
        "201":
          description: Match created
  /api/matches/tournament/{tournamentId}:
    get:
      summary: Get matches by tournament
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of matches
  /api/matches/{id}:
    get:
      summary: Get match by ID
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Match details
    put:
      summary: Update match
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Match updated
    delete:
      summary: Delete match
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Match deleted
  /api/matches/{id}/start:
    post:
      summary: Start match
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Match started
  /api/matches/{id}/end:
    post:
      summary: End match
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Match ended
  /api/matches/{id}/score:
    post:
      summary: Update match score
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Score updated
  /api/matches/{id}/officials:
    post:
      summary: Add match official
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Official added
    get:
      summary: Get match officials
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of officials
  /api/matches/{id}/officials/{userId}:
    delete:
      summary: Remove match official
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Official removed
  /api/matches/{id}/with-period:
    get:
      summary: Get match with period info
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Match with current period and substitution status
  /api/matches/{id}/period/transition:
    post:
      summary: Transition match period
      tags:
        - Matches
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum:
                    - start_first_half
                    - end_first_half
                    - start_second_half
                    - end_second_half
                    - start_extra_time
                    - end_extra_time_first
                    - start_extra_time_second
                    - end_extra_time
                    - start_penalties
                    - end_match
                    - abandon
                    - suspend
                    - resume
                injury_time:
                  type: integer
                winner_id:
                  type: string
                reason:
                  type: string
      responses:
        "200":
          description: Period transitioned
  /api/matches/{id}/abandon:
    post:
      summary: Abandon match
      tags:
        - Matches
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Match abandoned
  /api/matches/{id}/suspend:
    post:
      summary: Suspend match
      tags:
        - Matches
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Match suspended
  /api/matches/{id}/resume:
    post:
      summary: Resume suspended match
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Match resumed
  /api/matches/{id}/postpone:
    post:
      summary: Postpone match
      tags:
        - Matches
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - new_date
                - reason
              properties:
                new_date:
                  type: string
                reason:
                  type: string
      responses:
        "200":
          description: Match postponed
...
  /api/notifications:
    get:
      summary: Get user notifications
      tags:
        - Notifications
      security:
        - bearerAuth: []
      parameters:
        - name: unread
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of notifications
  /api/notifications/unread-count:
    get:
      summary: Get unread notification count
      tags:
        - Notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
  /api/notifications/read-all:
    post:
      summary: Mark all as read
      tags:
        - Notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All marked as read
  /api/notifications/{id}/read:
    post:
      summary: Mark notification as read
      tags:
        - Notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notification marked as read
  /api/notifications/{id}:
    delete:
      summary: Delete notification
      tags:
        - Notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notification deleted
...
  /api/visualization/annotations/match/{matchId}:
    get:
      summary: Get match annotations
      tags:
        - Visualization
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of annotations
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. Get it from POST /api/auth/login. Pass as:
        Authorization: Bearer <token>"
...
security:
  - bearerAuth: []
...